create updateUrlAfterDelete util
